ref(release_health): Rename module to be consistent with function name conventions
