+++ mypy.ini
@@ -44,7 +44,7 @@ files = src/sentry/api/bases/external_actor.py,
         src/sentry/models/projectoption.py,
         src/sentry/models/useroption.py,
         src/sentry/notifications/**/*.py,
-        src/sentry/releasehealth/**/*.py,
+        src/sentry/release_health/**/*.py,
         src/sentry/sentry_metrics/**/*.py,
         src/sentry/shared_integrations/constants.py,
         src/sentry/snuba/outcomes.py,
