+++ src/sentry/models/release.py
@@ -1049,7 +1049,7 @@ class Release(Model):
         """Deletes a release if possible or raises a `UnsafeReleaseDeletion`
         exception.
         """
-        from sentry import releasehealth
+        from sentry import release_health
         from sentry.models import Group, ReleaseFile
 
         # we don't want to remove the first_release metadata on the Group, and
@@ -1063,7 +1063,7 @@ class Release(Model):
         # We would need to be able to delete this data from snuba which we
         # can't do yet.
         project_ids = list(self.projects.values_list("id").all())
-        if releasehealth.check_has_health_data([(p[0], self.version) for p in project_ids]):
+        if release_health.check_has_health_data([(p[0], self.version) for p in project_ids]):
             raise UnsafeReleaseDeletion(ERR_RELEASE_HEALTH_DATA)
 
         # TODO(dcramer): this needs to happen in the queue as it could be a long
