+++ src/sentry/release_health/metrics.py
@@ -7,7 +7,7 @@ from snuba_sdk.expressions import Granularity
 from snuba_sdk.query import SelectableExpression
 
 from sentry.models.project import Project
-from sentry.releasehealth.base import (
+from sentry.release_health.base import (
     CurrentAndPreviousCrashFreeRates,
     EnvironmentName,
     OrganizationId,
@@ -140,7 +140,7 @@ class MetricsReleaseHealthBackend(ReleaseHealthBackend):
         )
 
         count_data = raw_snql_query(
-            count_query, referrer="releasehealth.metrics.get_crash_free_data", use_cache=False
+            count_query, referrer="release_health.metrics.get_crash_free_data", use_cache=False
         )["data"]
 
         for row in count_data:
@@ -301,18 +301,18 @@ class MetricsReleaseHealthBackend(ReleaseHealthBackend):
 
         # Count of sessions/users for given list of environments and timerange, per-project
         sessions_per_project: Dict[int, int] = _count_sessions(
-            total=True, referrer="releasehealth.metrics.get_release_adoption.total_sessions"
+            total=True, referrer="release_health.metrics.get_release_adoption.total_sessions"
         )
         users_per_project: Dict[int, int] = _count_users(
-            total=True, referrer="releasehealth.metrics.get_release_adoption.total_users"
+            total=True, referrer="release_health.metrics.get_release_adoption.total_users"
         )
 
         # Count of sessions/users for given list of environments and timerange AND GIVEN RELEASES, per-project
         sessions_per_release: Dict[Tuple[int, int], int] = _count_sessions(
-            total=False, referrer="releasehealth.metrics.get_release_adoption.releases_sessions"
+            total=False, referrer="release_health.metrics.get_release_adoption.releases_sessions"
         )
         users_per_release: Dict[Tuple[int, int], int] = _count_users(
-            total=False, referrer="releasehealth.metrics.get_release_adoption.releases_users"
+            total=False, referrer="release_health.metrics.get_release_adoption.releases_users"
         )
 
         rv = {}
@@ -420,7 +420,7 @@ class MetricsReleaseHealthBackend(ReleaseHealthBackend):
         )
 
         result = raw_snql_query(
-            query, referrer="releasehealth.metrics.check_has_health_data", use_cache=False
+            query, referrer="release_health.metrics.check_has_health_data", use_cache=False
         )
 
         return {extract_raw_info(raw) for raw in result["data"]}
